# swagger.yaml
openapi: "3.0.2"
info:
  title: 2FA Demo
  description: API for 2FA Demo
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
paths:
...
  /user/login:
    get:
      tags:
        - "User"
      summary: Get Login Page
      description: |
        This operation is done to get an html response,
        which render a login form
      responses:
        "200":
          description: OK
          content:
            text/html:
              example: <html>Login Page</html>
    post:
      tags:
        - "User"
      summary: Post Login Request
      description: |
        This operation is done to send login credentials which
        are then validated before logging in the user
      requestBody:
        required: true
        description: Form data containing login credentials
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LoginUser"
      responses:
        "200":
          description: OK
          content:
            text/html:
              example: <html>Login Page</html>
...
  /2fa/register:
    get:
      tags:
        - "2FA"
      summary: Get 2FA Registration Page
      description: |
        This operation is done to get an html response,
        which render a form to input token
      responses:
        "200":
          description: OK
          content:
            text/html:
              example: <html>2FA Registration Page</html>
    post:
      tags:
        - "2FA"
      summary: Send Token for Registration
      description: |
        This operation is done to send token which is then 
        validated before enabling 2FA for the user
      requestBody:
        required: true
        description: Form data containing Registration token
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/EnterToken"
      responses:
        "200":
          description: OK
          content:
            text/html:
              example: <html>2FA Registration Page</html>
...
components:
  schemas:
    LoginUser:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
    EnterToken:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...